+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Version 0.7.6
 
+- Added option to order raster spike trains by multipe tags (eg. ['pop', 'y'])
+
 - Enable providing a 2D list of spike times to VecStim populations
 
 - Added exception handling to analysis functions
@@ -12,11 +14,15 @@
 
 - Made CVode object part of sim and simplified cvode calls
 
+- Added __rename__() method to dict class
+
 - Fixed bug loading from compactConnFormat
 
 - Fixed bug in IClamp: 'delay' -> 'del'
 
 - Fixed bug calculating min convergece and divergence
+
+- Fixed bug in plotRatePSD colors
 
 
 # Version 0.7.5